discord_bot/src/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
Deprecated==1.2.13
discord.py==2.0.1
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
packaging==21.3
pyparsing==3.0.9
redis==4.3.4
requests==2.28.1
urllib3==1.26.12
wrapt==1.14.1
yarl==1.8.1
